Fixing the issue with saving UserFavorites with the favorite movies. Created Many to Many mapping Renamed MovieInfo to Movie.

+++ src/main/java/com/techreturners/movieapi/repository/UserFavoriteRepository.java
@@ -3,6 +3,6 @@ package com.techreturners.movieapi.repository;
 import com.techreturners.movieapi.model.UserFavorite;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserFavoriteRepository  extends CrudRepository<UserFavorite, Long> {
+public interface UserFavoriteRepository  extends CrudRepository<UserFavorite, Integer> {
 
 }
